DEV-2352
git-svn-id: https://svn.terracotta.org/repo/tc/dso/trunk@12152 7fc7bbf3-cf45-46d4-be06-341739edd864

+++ code/base/dso-common/tests.base/com/tc/net/protocol/tcm/TestMessageChannel.java
@@ -34,7 +34,7 @@ public class TestMessageChannel implements MessageChannel {
 
   public NodeID getLocalNodeID() {
     if (source == ClientID.NULL_ID) {
-      source = new ClientID(getChannelID());
+      source = new ClientID(getChannelID().toLong());
     }
     return source;
   }
