+++ code/base/dso-l2/tests.base/com/tc/object/net/TestDSOChannelManager.java
@@ -80,6 +80,6 @@ public class TestDSOChannelManager implements DSOChannelManager {
   }
 
   public ClientID getClientIDFor(ChannelID channelID) {
-    return new ClientID(channelID);
+    return new ClientID(channelID.toLong());
   }
 }
