+++ code/base/dso-l2/tests.base/com/tc/objectserver/lockmanager/api/NullChannelManager.java
@@ -64,7 +64,7 @@ public class NullChannelManager implements DSOChannelManager {
   }
 
   public ClientID getClientIDFor(ChannelID channelID) {
-    return new ClientID(channelID);
+    return new ClientID(channelID.toLong());
   }
 
 }
